requirements.txt.txt
python-telegram-bot==20.7
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
yfinance==0.2.18
requests==2.31.0